fix(cli): wire OUTPUTS persistence into auto-execution path
storeStepOutputs was only called from executeTransition (manual rd pass).
Auto-executed command steps and drained substep completions went through
applyResultTransition without ever storing OUTPUTS, silently losing them.

- Extract storeStepOutputs into helpers/step-outputs.ts (shared)
- Add cwd to ApplyResultTransitionArgs and DrainResolvedCompletionsArgs
- Call storeStepOutputs in applyResultTransition on PASS, guarded by a
  substep-context check so parent OUTPUTS are stored only when the step
  itself advances (not for each individual substep completion)
- Propagate cwd through drainResolvedCompletions and all call sites in
  runExecutionLoop and transitions.ts
